requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
cbor==1.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
google-auth==2.19.1
idna==3.4
iso8601==1.1.0
kopf==1.36.1
kubernetes==26.1.0
multidict==6.0.4
oauthlib==3.2.2
oisp==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-dateutil==2.8.2
python-json-logger==2.0.7
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
termcolor==2.3.0
typing-extensions==4.6.3
urllib3==2.0.2
websocket-client==1.5.2
yarl==1.9.2
